Configure spec helpers
test that 'helpers.rb' is available;
include Helpers for all spec files;
nest 'before' actions under 'feature' block
to rectify errors produced from improper nesting syntax;
