+++ spec/features/guest_user_searches_for_lodging_spec.rb
@@ -8,7 +8,9 @@ feature 'Guest user searches for lodging' do
     page.execute_script %Q{ $("a.ui-state-default:contains('15')").trigger("click") } # click on day 15
     page.execute_script %Q{ $('#to').trigger("focus") } # activate datetime picker
     page.execute_script %Q{ $("a.ui-state-default:contains('16')").trigger("click") } # click on day 16
+
     click_button 'Search'
+
     expect(current_path).to eq(listings_path)
     expect(page).to have_content('Asia')
   end
