+++ spec/features/guest_user_sees_list_of_lodging_options_spec.rb
@@ -1,19 +1,19 @@
 require 'rails_helper'
 
-before :each do
-  create_listings
-  create_locations
-
-  visit root_path
-  select('Asia', from: 'Location')
-  page.execute_script %Q{ $('#from').trigger("focus") } # activate datetime picker
-  page.execute_script %Q{ $("a.ui-state-default:contains('15')").trigger("click") } # click on day 15click_on 'Search'
-  page.execute_script %Q{ $('#to').trigger("focus") } # activate datetime picker
-  page.execute_script %Q{ $("a.ui-state-default:contains('16')").trigger("click") } # click on day 15click_on 'Search'
-  click_button 'Search'
-end
-
 feature 'Guest user sees a list of lodging options' do
+  before :each do
+    create_listings
+    create_locations
+
+    visit root_path
+    select('Asia', from: 'Location')
+    page.execute_script %Q{ $('#from').trigger("focus") } # activate datetime picker
+    page.execute_script %Q{ $("a.ui-state-default:contains('15')").trigger("click") } # click on day 15click_on 'Search'
+    page.execute_script %Q{ $('#to').trigger("focus") } # activate datetime picker
+    page.execute_script %Q{ $("a.ui-state-default:contains('16')").trigger("click") } # click on day 15click_on 'Search'
+    click_button 'Search'
+  end
+
   scenario 'on listings index, not logged in' do
     expect(page).to have_content('Tiki Hut 0')
     expect(page).to have_content('Tiki Hut 1')
