--- spec/helpers/locations.rb
@@ -1,17 +1,0 @@
-module LocationsHelper
-  def create_locations
-    places = [
-      'North America',
-      'South America',
-      'Europe',
-      'Africa',
-      'Asia',
-      'Australia',
-      'Antarctica'
-    ]
-
-    places.each do |place|
-      Location.create!(continent: place)
-    end
-  end
-end
