+++ spec/include_module_spec.rb
@@ -1,0 +1,5 @@
+RSpec.describe "test group" do
+  it "has access to the helper methods defined in the module" do
+    expect(help).to be(:available)
+  end
+end
