+++ spec/rails_helper.rb
@@ -4,8 +4,7 @@ require File.expand_path('../../config/environment', __FILE__)
 abort("The Rails environment is running in production mode!") if Rails.env.production?
 require 'spec_helper'
 require 'rspec/rails'
-require './helpers'
-include ListingsHelper
+require 'helpers'
 
 ActiveRecord::Migration.maintain_test_schema!
 
@@ -14,10 +13,4 @@ RSpec.configure do |config|
   config.use_transactional_fixtures = false
   config.infer_spec_type_from_file_location!
   config.include Helpers
-end
-
-RSpec.describe "create_listings is available" do
-  it "has access to the helper methods defined in the module" do
-    expect(create_listings).to be(:available)
-  end
 end
